% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adjoint.R
\name{adjoint}
\alias{adjoint}
\title{Adjugate or classical adjoint of a square matrix}
\usage{
adjoint(x)
}
\description{
The adjugate or classical adjoint of a square matrix is the transpose of its cofactor matrix.
It is also occasionally known as adjunct matrix,[ though this nomenclature appears to have decreased in usage.
}
