% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_bind.R
\name{cbind}
\alias{cbind}
\alias{rbind}
\title{Combine polynial matrices by rows or coluns}
\usage{
cbind(..., deparse.level = 1)

rbind(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalzed) vectors or mmatrices.
If any of objects are polynomail matrix}

\item{deparse.level}{details in base function, polynomial matrices doesn't use it}
}
\value{
if at least one argument is a polynomail matrix,
the result will be combined polynomial matrix.
Overwise, base package implementatioon \code{\link[base:cbind]{base::cbind()}} or \code{\link[base:cbind]{base::rbind()}}
will be called.
}
\description{
Combine polynial matrices by rows or coluns
}
\section{Functions}{
\itemize{
\item \code{rbind}: row based bind
}}

\seealso{
\code{\link[base:cbind]{base::cbind()}}
}
