% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_charpolynom.R, R/method_coef.R
\name{charpolynom}
\alias{charpolynom}
\alias{charpolynom,matrix-method}
\alias{charpolynom,polyMatrix-method}
\alias{[[,polyMatrixCharPolynomail,ANY-method}
\title{Characteristic polynomial of matrix}
\usage{
charpolynom(x)

\S4method{charpolynom}{matrix}(x)

\S4method{charpolynom}{polyMatrix}(x)

\S4method{[[}{polyMatrixCharPolynomail,ANY}(x, i)
}
\value{
When the input is a numerical matrix of \code{matrix} class
then the value is a \code{polynomial} object.

When the input is a \code{polyMatrix} object
then a value is \code{polyMatrixCharClass} class object,
}
\description{
Characteristic polynomial of matrix
}
\details{
The characteristic polynom of a polynomial matrix is a polynom with polynomial coefficients.
}
\section{Methods (by class)}{
\itemize{
\item \code{matrix}: for numerical matrix it is a polynomial with numerical coefficients

\item \code{polyMatrix}: for polynomial matrix it has polynomial coefficients

\item \code{x = polyMatrixCharPolynomail,i = ANY}: get polynomial coefficient of characteristic
}}

\examples{

# numerical matrices
m <- matrix(c(2, 1,
             -1, 0), 2, 2, byrow=TRUE)
charpolynom(m)
}
\seealso{
\link{polyMatrixCharClass}
}
