% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adjoint.R
\name{cofactor}
\alias{cofactor}
\title{Cofactor of matrix}
\usage{
cofactor(x, r, c)
}
\arguments{
\item{x}{an matrix}

\item{r, c}{the row and column}
}
\value{
cofactor which is number or polynomial
}
\description{
Cofactor of matrix
}
\seealso{
\code{\link[=adjoint]{adjoint()}}
}
