% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_gcd.R, R/method_lcm.R
\name{GCD}
\alias{GCD}
\alias{GCD,polyMatrix-method}
\alias{LCM,polyMatrix-method}
\title{GCD and LCM for polynomial matrices}
\usage{
GCD(...)

\S4method{GCD}{polyMatrix}(...)

\S4method{LCM}{polyMatrix}(...)
}
\description{
The greatest common divisor and least common multiple of polynomials or polynomial matrices.
}
\section{Methods (by class)}{
\itemize{
\item \code{polyMatrix}: the greatest common divisor of all elements of the polynomial matrice

\item \code{polyMatrix}: the least common multiple of polynomial matrices
}}

\examples{

# GCD of polynomial matrix

GCD(parse.polyMatrix(
 "  1 - x, 1 - x^2, 1 + 2*x + x^2",
 "x - x^2,   1 + x, 1 - 2*x + x^2"
))  ## 1

# LCM of polynomial matrix
LCM(parse.polyMatrix(
 "  1 - x, 1 - x^2, 1 + 2*x + x^2",
 "x - x^2,   1 + x, 1 - 2*x + x^2"
))  ## 0.25*x - 0.5*x^3 + 0.25*x^5
}
\seealso{
polynomial implementation \code{\link[polynom:gcd]{polynom::GCD()}}
}
