% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_polyMatrix.R
\name{is.polyMatrix}
\alias{is.polyMatrix}
\title{Check if object is polyMatrix}
\usage{
is.polyMatrix(x)
}
\arguments{
\item{x}{an \R object}
}
\value{
TRUE if object is a polonial matrix
}
\description{
Check if object is polyMatrix
}
\examples{
is.polyMatrix(c(1, 2, 3))
is.polyMatrix(polyMatrix(0, 2, 2))

}
