% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_det.R
\name{minor}
\alias{minor}
\title{Minor of matrix item}
\usage{
minor(x, r, c)
}
\arguments{
\item{x}{a matrix}

\item{r, c}{row and column}
}
\description{
A minor of a matrix A is the determinant of some smaller square matrix,
cut down from A by removing one or more of its rows and columns.
Minors obtained by removing just one row and one column from square matrices (first minors).
}
