% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_interpolation.R
\name{newton}
\alias{newton}
\title{Build matrix of polynimal decomposition using Newton interpolation in Newton bais:
(x-x_0), (x - x_0) * (x x_1)}
\usage{
newton(C, points)
}
\arguments{
\item{C}{Matrix of values of polinomials in columns}

\item{points}{point in which the values of polynomials were got}
}
\value{
Matrix of coefficients in columns (from higher degree to lower)
}
\description{
Build matrix of polynimal decomposition using Newton interpolation in Newton bais:
(x-x_0), (x - x_0) * (x x_1)
}
