% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse.polynomial}
\alias{parse.polynomial}
\title{Parse polynomial from string}
\usage{
parse.polynomial(s, var = "x")
}
\arguments{
\item{s}{an string for parsing}

\item{var}{an variable name}
}
\value{
new polynomial as \code{polynom::polynomial} object
}
\description{
Parse string representation of polynomial into a polynomial object.
}
\seealso{
\code{\link[=parse.polyMatrix]{parse.polyMatrix()}}
}
