\name{pMatrix-package}
\alias{pMatrix-package}
\docType{package}
\title{\packageTitle{pMatrix}}
\description{\packageDescription{pMatrix}}
\details{
  \tabular{ll}{
    Package: \tab pMatrix\cr
    Type: \tab Package\cr
    Version: \tab 0.0.05\cr
    Date: \tab 2016-06-04\cr
    License: \tab GPL-3 \cr
}
Infrastructure for manipulation polynomial matrices.
}
\author{
\packageAuthor{pMatrix}
\cr
Maintainer: \packageMaintainer{pMatrix}
}
\references{
See other polynomial packages: 
\code{polynom}, 
\code{PolynomF},
\code{MonoPoly},
\code{multipol},
\code{mpoly},
\code{orthopolynom}.
}
\examples{

# the  c("pMarray", "pMatrix") class structure
str(pMgen.a())

# the  c("pMbroad", "pMatrix") class structure
str(pMgen.b())

# the  c("pMcells", "pMatrix") class structure
str(pMgen.c())

# the  c("pMdlist", "pMatrix") class structure
str(pMgen.d())

}
\keyword{ package }


