\name{pMkmm}
\alias{pMkmm}
\title{The Kronecker indices of a polynomial matrix}
\description{ ... }

\usage{pMkmm(pm) }

\arguments{\item{pm}{A polynomial matrix}}

\details{more details than the description above }

\value{A vector of the Kronecker indices}

\note{further notes }

\author{Manel Boukhris [tun]}

\references{references to the literature }

\seealso{See Also as \code{\link{polyMatrix-package}}}

\examples{
polyMgen.a()
}

\keyword{math}

