% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_polyMatrix.R, R/method_coef.R,
%   R/method_dim.R, R/method_round.R, R/method_show.R
\docType{class}
\name{polyMatrix-class}
\alias{polyMatrix-class}
\alias{polyMatrixClass}
\alias{[[,polyMatrix,numeric-method}
\alias{nrow,polynomial-method}
\alias{ncol,polynomial-method}
\alias{dim,polyMatrix-method}
\alias{round,polyMatrix-method}
\alias{show,polyMatrix-method}
\title{A class to represent matrix of polinomials}
\usage{
\S4method{[[}{polyMatrix,numeric}(x, i)

\S4method{nrow}{polynomial}(x)

\S4method{ncol}{polynomial}(x)

\S4method{dim}{polyMatrix}(x)

\S4method{round}{polyMatrix}(x, digits = 0)

\S4method{show}{polyMatrix}(object)
}
\description{
A class to represent matrix of polinomials
}
\section{Methods (by generic)}{
\itemize{
\item \code{[[}: get coefficient matrix by degree

\item \code{nrow}: number of rows of a polynomial matrix

\item \code{ncol}: number of column of a polynomial matrix

\item \code{dim}: dimension of a polynomial matrix

\item \code{round}: round of polynomial matrix is rounding of polynomial coefficients

\item \code{show}: prints out a text representation of a polynomial matrix
}}

\section{Slots}{

\describe{
\item{\code{coef}}{A matrix of coefficients which are joined into one matrix from lower degree to higher}

\item{\code{ncol}}{Actual number of columns in the polynomial matrix}
}}

\examples{

# create a new polynomial matrix by parsing strings
pm <- parse.polyMatrix(
     "x; 1 + x^2; 3 x - x^2",
     "1; 1 + x^3; - x + x^3"
)

# get coefficient matrix for degree 0
pm[[0]]
##      [,1] [,2] [,3]
## [1,]    0    1    0
## [2 ]    1    1    0
# get coefficient matrix for degree 1
pm[[1]]
##      [,1] [,2] [,3]
## [1,]    1    0    3
## [2 ]    0    0   -1


# dimensions
nrow(pm) ## 2

ncol(pm) ## 3

dim(pm) ## [1] 2 3

# round
round(parse.polyMatrix(
  "      1.0001 - x,            1 - x^2, 1 + 2.0003*x + x^2",
  "0.0001 + x - x^2, 1 + x + 0.0001 x^2, 1 - 2*x + x^2"
))
##           [,1]      [,2]           [,3]
## [1,]     1 - x   1 - x^2   1 + 2x + x^2
## [2,]   x - x^2     1 + x   1 - 2x + x^2

# print out a polynomial matrix
show(parse.polyMatrix(
  "      1.0001 - x,          1 - x^2, 1 + 2.0003*x + x^2",
  "0.0001 + x - x^2,            1 + x, 1 - 2*x + x^2",
  "        12.3 x^3,  2 + 3.5 x + x^4, -0.7 + 1.6e-3 x^3"
))
##                   [,1]             [,2]                [,3]
## [1,]        1.0001 - x          1 - x^2   1 + 2.0003x + x^2
## [2,]   1e-04 + x - x^2            1 + x        1 - 2x + x^2
## [3,]           12.3x^3   2 + 3.5x + x^4    -0.7 + 0.0016x^3
}
