% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_matrix_mult.R
\name{\%*\%,polyMatrix,polyMatrix-method}
\alias{\%*\%,polyMatrix,polyMatrix-method}
\alias{\%*\%,polyMatrix,matrix-method}
\alias{\%*\%,matrix,polyMatrix-method}
\title{Matrix multiplication}
\usage{
\S4method{\%*\%}{polyMatrix,polyMatrix}(x, y)

\S4method{\%*\%}{polyMatrix,matrix}(x, y)

\S4method{\%*\%}{matrix,polyMatrix}(x, y)
}
\arguments{
\item{x, y}{first and second operands}
}
\description{
Matrix multiplication accepts both polynomial and numerical matrices.
}
