% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_subscript.R
\name{[,polyMatrix,missing,missing,missing-method}
\alias{[,polyMatrix,missing,missing,missing-method}
\alias{[,polyMatrix,missing,ANY,missing-method}
\alias{[,polyMatrix,ANY,missing,missing-method}
\alias{[,polyMatrix,logical,logical,missing-method}
\alias{[,polyMatrix,logical,numeric,missing-method}
\alias{[,polyMatrix,numeric,logical,missing-method}
\alias{[,polyMatrix,numeric,numeric,missing-method}
\alias{[<-,polyMatrix,missing,missing,ANY-method}
\alias{[<-,polyMatrix,missing,ANY,ANY-method}
\alias{[<-,polyMatrix,ANY,missing,ANY-method}
\alias{[<-,polyMatrix,numeric,numeric,numeric-method}
\alias{[<-,polyMatrix,numeric,numeric,matrix-method}
\alias{[<-,polyMatrix,numeric,numeric,polynomial-method}
\alias{[<-,polyMatrix,numeric,numeric,polyMatrix-method}
\title{Extract or Replace Parts of a polynomial matrix}
\usage{
\S4method{[}{polyMatrix,missing,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{polyMatrix,missing,ANY,missing}(x, i, j)

\S4method{[}{polyMatrix,ANY,missing,missing}(x, i, j)

\S4method{[}{polyMatrix,logical,logical,missing}(x, i, j)

\S4method{[}{polyMatrix,logical,numeric,missing}(x, i, j)

\S4method{[}{polyMatrix,numeric,logical,missing}(x, i, j)

\S4method{[}{polyMatrix,numeric,numeric,missing}(x, i, j)

\S4method{[}{polyMatrix,missing,missing,ANY}(x, i, j) <- value

\S4method{[}{polyMatrix,missing,ANY,ANY}(x, i, j) <- value

\S4method{[}{polyMatrix,ANY,missing,ANY}(x, i, j) <- value

\S4method{[}{polyMatrix,numeric,numeric,numeric}(x, i, j) <- value

\S4method{[}{polyMatrix,numeric,numeric,matrix}(x, i, j) <- value

\S4method{[}{polyMatrix,numeric,numeric,polynomial}(x, i, j) <- value

\S4method{[}{polyMatrix,numeric,numeric,polyMatrix}(x, i, j) <- value
}
\arguments{
\item{x}{a polynomial matrix}

\item{i}{row indeces}

\item{j}{column indeces}

\item{...}{unused}

\item{drop}{unused}

\item{value}{new value}
}
\description{
Extract or Replace Parts of a polynomial matrix
}
\section{Functions}{
\itemize{
\item \code{[,polyMatrix,missing,ANY,missing-method}: get columns

\item \code{[,polyMatrix,ANY,missing,missing-method}: gets rows

\item \code{[,polyMatrix,logical,logical,missing-method}: gets by logical index

\item \code{[,polyMatrix,logical,numeric,missing-method}: gets by logical index and numerical indices

\item \code{[,polyMatrix,numeric,logical,missing-method}: gets by logical index and numerical indices

\item \code{[,polyMatrix,numeric,numeric,missing-method}: gets by row and column indices

\item \code{[<-,polyMatrix,missing,missing,ANY-method}: replace o matrix by a new one

\item \code{[<-,polyMatrix,missing,ANY,ANY-method}: assigns rows

\item \code{[<-,polyMatrix,ANY,missing,ANY-method}: assigns columns

\item \code{[<-,polyMatrix,numeric,numeric,numeric-method}: replace part of matrix by one number

\item \code{[<-,polyMatrix,numeric,numeric,matrix-method}: replace part of matrix by another numerical matrix.
Size of the new matrix should be same as replaced part

\item \code{[<-,polyMatrix,numeric,numeric,polynomial-method}: replace part of matrix by one polynomail

\item \code{[<-,polyMatrix,numeric,numeric,polyMatrix-method}: replace part of matrix by another polunomial matrix.
Size of the new matrix should be same as replaced part
}}

