% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{polyMatrix.apply}
\alias{polyMatrix.apply}
\title{Apply for polynomial matrix}
\usage{
polyMatrix.apply(x, f)
}
\arguments{
\item{x}{an polynomial matrix}

\item{f}{an function with only one argument}
}
\description{
Apply function to each element of matrix
}
