% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_polyMatrix.R
\name{polyMatrix}
\alias{polyMatrix}
\title{Create polyMatrix object}
\usage{
polyMatrix(data, nrow, ncol, degree)
}
\arguments{
\item{data}{A matrix containing matrices of coefficients
or a number or a polynomial}

\item{nrow}{The numer of rows of a polynomial matrix. Must be postive.
If data is a matrix, the default value is the number of rows of matrix \code{data}.
In other cases it is a required parameter.}

\item{ncol}{A number of columns of a polynomial matrix. Must be positive.
If data is a matrix, the default value is the number of columns of matrix \code{data}.
In other cases it is a required parameter.}

\item{degree}{Degree of polynomials in the coefficient matrix. Must be zero or positive.
If data is polynomial, degree can be evaluated automatcal.
In other case, default value is 0.}
}
\value{
new polynomial matrix of polyMatrix class
}
\description{
This function will create a polynomial object from
coefficient matrix or signle value
}
\details{
A coefficient matrix is a matrix which contains
matrices of coefficients starting from lower degree to higher ones,
side-by-side
}
