% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_polyMatrixCharPolynomial.R
\docType{class}
\name{polyMatrixCharPolynomail-class}
\alias{polyMatrixCharPolynomail-class}
\alias{polyMatrixCharClass}
\title{A class to repesent characteristic polynomial of polynomial matrix}
\description{
Characteristic polynomial of polynomial matrix is polynomial with polynomial coefficients
}
