% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_polyMatrixCharPolynomial.R
\docType{class}
\name{polyMatrixCharPolynomial-class}
\alias{polyMatrixCharPolynomial-class}
\alias{polyMatrixCharClass}
\title{A class to repesent characteristic polynomial of a polynomial matrix}
\description{
Characteristic polynomial of a polynomial matrix is a polynomial with polynomial coefficients
}
