% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_euclidean.R
\name{triang.euclidean}
\alias{triang.euclidean}
\title{Triangularization of a polynomial matrix by Euclidean division method}
\usage{
triang.euclidean(pm, eps = ZERO_EPS, iteration_limit = nrow(pm) * 100)
}
\arguments{
\item{pm}{matrix to triangularize}

\item{eps}{threshold of non zero coefficients}
}
\value{
Upper tringular matrix \code{R}, transformation matrix \code{Q} and iverse \code{INV_Q} of matrix \code{Q}
}
\description{
Triangularization of a polynomial matrix by Euclidean division method
}
\details{
The method use the for polynomials extended Euclidean algorithm.
}
\seealso{
\code{\link[=triang_Sylvester]{triang_Sylvester()}}
}
