% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_interpolation.R
\name{triang_Interpolation}
\alias{triang_Interpolation}
\title{Triangularization of a polynomial matrix by interpolation method}
\usage{
triang_Interpolation(
  pm,
  point_vector,
  round_digits = 5,
  eps = .Machine$double.eps^0.5
)
}
\arguments{
\item{pm}{source polynimial matrix}

\item{point_vector}{vector of interpolation points}

\item{round_digits}{we will try to round result on each step}

\item{eps}{calculation zero errors}
}
\value{
Tranfortmaiton matrix
}
\description{
The parameters \code{point_vector}, \code{round_digits} can significantly affect the result.
}
\details{
Default value of `eps`` usually is enought to determintate real zeros.

In a polynomial matrix the head elements are the first non-zero polynomials of columns.
The sequence of row indices of this head elements form the shape of the polynomial matrix.
A polynomial matrix is in left-lower triangular form, if this sequence is monoton increasing.

This method offers a solution of the triangulrization by the Interpolation method,
described in the article of Labhalla-Lombardi-Marlin (1996).
}
