% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_utils.R
\name{zero_lead_hyp_rows}
\alias{zero_lead_hyp_rows}
\title{Get zero lead hyper rows of size sub_nrow of matrix M}
\usage{
zero_lead_hyp_rows(M, sub_nrow, esp = ZERO_EPS)
}
\arguments{
\item{M}{Numerical matrix}

\item{sub_nrow}{Size of hyper row}

\item{esp}{Machine epsilon to determinate zeros}
}
\value{
vector of idx of hyperrows, NaN for columns without zeros
}
\description{
Get zero lead hyper rows of size sub_nrow of matrix M
}
