% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_utils.R
\name{zero_lead_rows}
\alias{zero_lead_rows}
\title{Get zero lead rows of matrix M}
\usage{
zero_lead_rows(M, eps = ZERO_EPS)
}
\arguments{
\item{M}{Numerical matrix}

\item{eps}{Machine epsilon to determinate zeros}
}
\value{
vector of idx (length is equal to columm number), NULL in case of error
}
\description{
Get zero lead rows of matrix M
}
