% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adjoint.R
\name{adjoint}
\alias{adjoint}
\alias{adjoint,polyMatrix-method}
\title{Adjungate or classical adjoint of a square matrix}
\usage{
adjoint(x)

\S4method{adjoint}{polyMatrix}(x)
}
\arguments{
\item{x}{a matrix}
}
\description{
The adjungate or classical adjoint of a square matrix is the transpose of its cofactor matrix.
It is also occasionally known as adjunct matrix, though this nomenclature appears to have been decreased in usage.
}
\section{Methods (by class)}{
\itemize{
\item \code{polyMatrix}: adjungate of polynomial matrix DON'T UNDERSTAND!!!
}}

