% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adjoint.R
\name{adjoint}
\alias{adjoint}
\alias{adjoint,polyMatrix-method}
\title{Adjugate or classical adjoint of a square matrix}
\usage{
adjoint(x)

\S4method{adjoint}{polyMatrix}(x)
}
\arguments{
\item{x}{an matrix}
}
\description{
The adjugate or classical adjoint of a square matrix is the transpose of its cofactor matrix.
It is also occasionally known as adjunct matrix,[ though this nomenclature appears to have decreased in usage.
}
\section{Methods (by class)}{
\itemize{
\item \code{polyMatrix}: adjugate of polynomial matrix
}}

