% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_bind.R
\name{cbind}
\alias{cbind}
\alias{rbind}
\title{Combine polynomial matrices by rows or columns}
\usage{
cbind(..., deparse.level = 1)

rbind(..., deparse.level = 1)
}
\arguments{
\item{...}{(generalized) vectors or matrices.
If any of the objects is a polynomail matrix}

\item{deparse.level}{details in the base function, polynomial matrices doesn't use this argument}
}
\value{
if at least one argument is a polynomial matrix,
the result will be a combined polynomial matrix.
Otherwise, the base package implementation \code{\link[base:cbind]{base::cbind()}} or \code{\link[base:cbind]{base::rbind()}}
will be called.
}
\description{
Combine polynomial matrices by rows or columns
}
\section{Functions}{
\itemize{
\item \code{rbind}: row based bind  DON'T UNDERSTAND.. !!!
}}

\seealso{
\code{\link[base:cbind]{base::cbind()}}
}
