% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_adjoint.R
\name{cofactor}
\alias{cofactor}
\title{Cofactor of a matrix}
\usage{
cofactor(x, r, c)
}
\arguments{
\item{x}{a matrix}

\item{r, c}{the rows and columns}
}
\value{
cofactor which is a number or a polynomial
}
\description{
Cofactor of a matrix
}
\seealso{
\code{\link[=adjoint]{adjoint()}}
}
