\name{polyMatrix-package}
\alias{polyMatrix-package}
\docType{package}
\title{\packageTitle{polyMatrix}}
\description{\packageDescription{polyMatrix}}
\details{
  \tabular{ll}{
    Package: \tab polyMatrix\cr
    Type: \tab Package\cr
    Version: \tab 0.0.05\cr
    Date: \tab 2016-06-04\cr
    License: \tab GPL-3 \cr
}
Infrastructure for manipulation polynomial matrices.
}
\author{
\packageAuthor{polyMatrix}
\cr
Maintainer: \packageMaintainer{polyMatrix}
}
\references{
See other polynomial packages: 
\code{polynom}, 
\code{PolynomF},
\code{MonoPoly},
\code{multipol},
\code{mpoly},
\code{orthopolynom}.
}
\examples{

# the  c("polyMarray", "polyMatrix") class structure
str(polyMgen.a())

# the  c("polyMbroad", "polyMatrix") class structure
str(polyMgen.b())

# the  c("polyMcells", "polyMatrix") class structure
str(polyMgen.c())

# the  c("polyMdlist", "polyMatrix") class structure
str(polyMgen.d())

}
\keyword{ package }


