\name{pMkmm}
\alias{pMkmm}
\title{The Kronecker indices of a polynomial matrix}

\description{
Calculates the Kronecker indices of a polynomial matrix.
}

\usage{pMkmm(pm) }

\arguments{\item{pm}{a polynomial matrix}}

\value{A vector of the Kronecker indices.}

\note{The method doesn't work yet.}

%\references{references to the literature }

\seealso{\code{\link{polyMatrix-package}}}

\examples{
\donttest{
polyMgen.a()
} #% end of donttest
}

\keyword{math}

