% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triang_euclidean.R
\name{triang_Euclidean}
\alias{triang_Euclidean}
\title{Triangularization of a polynomial matrix by Euclidean division metho}
\usage{
triang_Euclidean(pm)
}
\arguments{
\item{pm}{matrix to triangularize}
}
\value{
Upper tringular matrix `T` and transformation matrix `U`.
}
\description{
Triangularization of a polynomial matrix by Euclidean division metho
}
\details{
The method use the for polynomilas extended Euclidean algorithm.

This method search a solution of the triangulrization by the method of Sylvester matrix,
descripted in the article Labhalla-Lombardi-Marlin (1996).
}
\references{
Thomas Kailaith:Linear Systems, Prentice-Hall, 1980, pp 373-376
}
\seealso{
triang_Sylvester
}
\author{
Nikolai Ryzhkov, \email{namezys@gmail.com}
}
